Add option to test if pwd has been reused
This breaking change closes #2 by including a new option (by default
turned on) to use the Have I Been Pwned (hibp) database using the open
REST API service, to check if the password has been exposed in a
previous breach. If it has, then it will be a POOR password.

If the option is enabled, but the hibp service is not available then
the report will still be returned, but an error explaining this issue
will be returned.

This brings a breaking change as the function will now return a promise
of the report, instead of the report itself. In addition, the options
have been renamed so they are positively named (enable...) for clarity.